requirements.txt
Django==1.5.2
South==0.8.2
argparse==1.2.1
boto==2.11.0
django-json-field==0.5.5
django-storages==1.1.8
python-dateutil==2.1
pytz==2013b
requests==1.2.3
six==1.4.1
wsgiref==0.1.2
